Move completion() from forms to data-model
The api package depends only on data-model, so it could not compute
form completion. completion(), its types and formDataToValues now live
in data-model with an injected per-field-type override resolver, and
@faims3/forms re-exports the same public API delegating to it.
Behavior unchanged.
